+++ backend/database/migrations/2018_12_30_160714_create_produtos_table.php
@@ -17,8 +17,8 @@ class CreateProdutosTable extends Migration
             $table->increments('id');
             $table->string('descricao',45);
             $table->string('detalhamento');
-            $table->float('preco_vista');
-            $table->float('preco_prazo');
+            $table->integer('preco_vista');
+            $table->integer('preco_prazo');
             $table->integer('codigo_barras');
             $table->integer('usuarios_id')->unsigned();
             $table->foreign('usuarios_id')->references('id')->on('usuarios');
